Restructure events modules
