+++ src/main/java/com/github/mkram17/bazaarutils/events/minecraft/ContainerLoadedEvent.java
@@ -1,4 +1,4 @@
-package com.github.mkram17.bazaarutils.events;
+package com.github.mkram17.bazaarutils.events.minecraft;
 
 import com.github.mkram17.bazaarutils.BazaarUtils;
 import com.github.mkram17.bazaarutils.utils.annotations.autoregistration.RunOnInit;
