+++ src/main/java/com/github/mkram17/bazaarutils/events/minecraft/ScreenChangeEvent.java
@@ -1,4 +1,4 @@
-package com.github.mkram17.bazaarutils.events;
+package com.github.mkram17.bazaarutils.events.minecraft;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
