+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/buttons/inputhelper/AmountHelpers.java
@@ -1,10 +1,10 @@
 package com.github.mkram17.bazaarutils.features.gui.buttons.inputhelper;
 
 import com.github.mkram17.bazaarutils.config.features.gui.ButtonsConfig;
-import com.github.mkram17.bazaarutils.events.ContainerLoadedEvent;
-import com.github.mkram17.bazaarutils.events.ReplaceItemEvent;
-import com.github.mkram17.bazaarutils.events.SlotClickEvent;
-import com.github.mkram17.bazaarutils.events.listener.BUListener;
+import com.github.mkram17.bazaarutils.events.minecraft.ContainerLoadedEvent;
+import com.github.mkram17.bazaarutils.events.minecraft.ReplaceItemEvent;
+import com.github.mkram17.bazaarutils.events.minecraft.SlotClickEvent;
+import com.github.mkram17.bazaarutils.events.BUListener;
 import com.github.mkram17.bazaarutils.utils.annotations.modules.Module;
 import com.github.mkram17.bazaarutils.utils.bazaar.SignInputHelper;
 import meteordevelopment.orbit.EventHandler;
