+++ src/main/java/com/github/mkram17/bazaarutils/mixin/MixinSignEditScreen.java
@@ -1,7 +1,7 @@
 package com.github.mkram17.bazaarutils.mixin;
 
 import com.github.mkram17.bazaarutils.BazaarUtils;
-import com.github.mkram17.bazaarutils.events.SignOpenEvent;
+import com.github.mkram17.bazaarutils.events.minecraft.SignOpenEvent;
 import net.minecraft.client.gui.screens.inventory.SignEditScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
