+++ src/main/java/com/github/mkram17/bazaarutils/utils/bazaar/data/BazaarDataManager.java
@@ -2,7 +2,7 @@ package com.github.mkram17.bazaarutils.utils.bazaar.data;
 
 import com.github.mkram17.bazaarutils.BazaarUtils;
 import com.github.mkram17.bazaarutils.utils.APIUtil;
-import com.github.mkram17.bazaarutils.events.BazaarDataUpdateEvent;
+import com.github.mkram17.bazaarutils.events.bazaar.BazaarDataUpdateEvent;
 import com.github.mkram17.bazaarutils.misc.NotificationType;
 import com.github.mkram17.bazaarutils.utils.PlayerActionUtil;
 import com.github.mkram17.bazaarutils.utils.Util;
